rapheye-fastapi/requirements.txt
fastapi
uvicorn[standard] # Incluye dependencias web estándar
python-multipart # Para subir archivos
Pillow
# Pix2tex y sus dependencias (asegúrate de que torch sea para CPU)
pix2tex @ git+https://github.com/lukas-blecher/LaTeX-OCR.git
torch>=1.7.1 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.8.2 --index-url https://download.pytorch.org/whl/cpu
torchaudio>=0.7.2 --index-url https://download.pytorch.org/whl/cpu
opencv-python-headless
albumentations>=1.3
numpy>=1.19
pyyaml>=5.4
tqdm>=4.62
scikit-image>=0.18
requests>=2.25
# Logging (ya incorporado, pero bueno tenerlo explícito si usaras config avanzada)
# python-dotenv # Si usaras variables de entorno desde .env localmente  